utils have been separated and views and viewmodels affected by AppConstants have been changed
