+++ CustomiTunes/Modules/SearchView/SearchView-ViewModel.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 @MainActor class SearchViewModel: ObservableObject {
-    let staticAppData = StaticAppData()
     @Published var searchResult: [SongData] = []
     @Published var searchKey: String = "" {
         didSet {
